Move Prezly Client one level higher
