+++ src/Sdk/index.ts
@@ -1,5 +1,3 @@
-export { Client as PrezlyClient, createClient as createPrezlyClient } from './Client';
-
 export * from './Accounts';
 export * from './Campaigns';
 export * from './CampaignRecipients';
@@ -12,4 +10,3 @@ export * from './NewsroomCategories';
 export * from './NewsroomContacts';
 export * from './NewsroomLanguages';
 export * from './Snippets';
-export * from './types';
