--- src/Sdk/types.ts
@@ -1,7 +1,0 @@
-import { HeadersMap } from '../http';
-
-export interface ClientOptions {
-    accessToken: string;
-    baseUrl?: string;
-    headers?: HeadersMap;
-}
