+++ src/index.ts
@@ -1,4 +1,6 @@
 export * from '@prezly/uploads';
 
+export { Client as PrezlyClient, createClient as createPrezlyClient } from './Client';
+
 export * from './Sdk';
 export * from './types';
